begin to port from e2b infra
There are four important components:
1. orchestrator
2. template-manager
3. api
4. envd

The envd can be used without few modifications and is running inside fc.
The template-manager has been simplified as a one-shot binary.
The left work is for orchestrator and api, in my design the orchestrator
and api might merge into one package and provide some simple interface.