Add a self-built code interpreter image.
The packages/fc-rootfs contains a docker image that can be dumped as
rootfs for FC, which contains a jupyter notebook that is compatiable
with sandbox_sdk.code_interpreter.

Furthermore, this docker image has fixed some issues in e2b (this
jupyter server is purely run under user, instead of root).