Enable host page cache sharing.
To solve the problems of multiple page cache on the host, I choose to
keep only one rootfs file on the host and make it read-only and share by
all sandboxes. Besides, to support writable inside the sandbox, I choose
to create an overlayfs in the VM. It also the solution from
firecracker-containerd. Now the configuration is:

1. Each sandbox is equipped two block devices. The first is read-only,
  containing the entire image and shared by all sandboxes. The second is
  a writable and private for each sandbox.
2. Includes a new init (called overlay-init), which will mount an overlay
  fs (with upper dir on the writable dev) and pivot_root.

The result is that all sandbox on the same machine will share page cache
on the host side. In the previous version, each sandbox will contain a
seperate page cache (for their rootfs image).