Add a new cli tools.
The sandbox will not be closed properly when the orchestrator has
crashed (or due to SIGKILL). Currently, the orchestrator will cleanup
the sandboxes in its signal handler.

In this commit, I add a new cli tool (which is a client to
orchestrator). It supports to:
1. list the sandboxes (including the valid ones and the orphan ones).
The orphan sandbox means that running in the OS but not maintained by
the current orchestrator (e.g., left by previous crashed orchestrator).
2. purge the sandbox. Note that purge is not identical to delete. Purge
means only cleanup the *orphan* sandboxes.

Probably in the future, we should try to persist the sandbox information
(e.g., BoltDB like Docker).