Make sure firecracker socket is ready.
There is a bug that happen rarely. The loadsnapshot request to
firecracker might failed due to "connection refused". As currently, we
only wait the socket exists (i.e., created by FC), but the sockets might
still not be ready (i.e., the FC has not fully initialized). In this
commit, the WaitForSocket() will return after sending a request
successfully.